chore: inline update pipeline
